resolve dtype mismatch

+++ medqa_steering/calibration/apply_ats.py
@@ -8,9 +8,12 @@ def load_ats(dim):
     head.eval()
     return head
 
+
 @torch.no_grad()
 def apply_ats(head, h, logits4):
-    # h: [1,d]; logits4: [1,4]
+    # Ensure dtype alignment (ATSHead is float32)
+    h = h.float()
+    logits4 = logits4.float()
     tau = head(h).clamp(min=1e-3, max=50.0)  # [1,1]
     z = logits4 / tau
     return z.softmax(dim=-1)
